adaptado componentes Login y Regster en Auth jsx
