+++ package.json
@@ -18,7 +18,6 @@
     "react-redux": "8.0.1",
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.0",
-    "react-toastify": "9.0.1",
     "redux": "4.2.0",
     "redux-thunk": "2.4.1",
     "sweetalert2": "11.4.14",
