+++ src/hooks/useAuthUser/useAuthUserForm.js
@@ -1,11 +1,10 @@
+import { useEffect } from "react"
+import { useLocation, useNavigate } from "react-router-dom"
 import { useFormik } from "formik"
 
-import { useEffect } from "react"
 import { useDispatch, useSelector } from "react-redux"
-import { useLocation, useNavigate } from "react-router-dom"
-
 import { loginUser, registerUser, resetUserNotification } from "redux/store/actions/userActions"
-import { useNotification } from "hooks/useNotification"
+
 import { alertMsg } from "utilities/utilAlert/utilAlertMsg"
 import { startValues, validateUserFormFields } from "utilities/utilAuthUser/utilUserFormsFields/utilUserFormsFields"
 
@@ -17,45 +16,14 @@ export const useUserForm = () => {
 
   const navigate = useNavigate()
 
-  const { loadingUser, user, error, success_request } = useSelector((state) => {
+  const { loadingUser, user, error, status_code, success_request } = useSelector((state) => {
     return state.userReducer
   })
   const dispatch = useDispatch()
 
-  const { showNotification } = useNotification()
-  // console.log(useNotification(error, success_request))
-
   const initialValues = startValues(pathName)
 
   const validationSchema = validateUserFormFields(pathName)
-
-  useEffect(() => {
-    if (error) {
-      console.log("TRUE ERRORRRRRR", error)
-      alertMsg({ title: "ERROR", text: `${error}`, icon: "error" })
-      dispatch(resetUserNotification())
-    } else if (success_request) {
-      console.log("estoy en EFECTOOO --->>", { success_request })
-      dispatch(resetUserNotification())
-
-      const possibleRoutes = {
-        1: "/",
-        2: "/login",
-        3: `/registered/${user.teamID}`,
-      }
-
-      let routeToNavigate = ""
-      if (pathName === "login") routeToNavigate = 1
-      else if (user.role === "Team Member") routeToNavigate = 2
-      else routeToNavigate = 3
-
-      // navigate(possibleRoutes[routeToNavigate])
-
-      console.log("user final -->", user)
-      console.log("routeToNavigate -->", routeToNavigate)
-      console.log("possibleRoutes[routeToNavigate] -->", possibleRoutes[routeToNavigate])
-    }
-  }, [user, error, success_request, dispatch, pathName, navigate])
 
   const onSubmit = () => {
     if (pathName === "login") dispatch(loginUser(formik.values))
@@ -63,6 +31,8 @@ export const useUserForm = () => {
       formik.values.teamID = !formik.values.teamID ? uuidv4() : formik.values.teamID
       dispatch(registerUser(formik.values))
     }
+
+    formik.resetForm()
   }
 
   const formik = useFormik({ initialValues, validationSchema, onSubmit })
@@ -76,5 +46,37 @@ export const useUserForm = () => {
   // change value switch - checked - not check
   const handleChangeSwitch = () => formik.setFieldValue("switch", !formik.values.switch)
 
-  return { formik, loadingUser, showNotification, pathName, handleChangeSwitch, handleChangeContinent }
+  useEffect(() => {
+    if (error) {
+      // console.log("TRUE ERRORRRRRR", error)
+      alertMsg({ title: "ERROR", text: `${error}`, icon: "error" })
+      dispatch(resetUserNotification())
+    } else if (success_request) {
+      // console.log("estoy en EFECTOOO --->>", { success_request })
+      dispatch(resetUserNotification())
+
+      const possibleRoutes = {
+        1: "/",
+        2: "/login",
+        3: `/registered/${user.teamID}`,
+      }
+
+      let routeToNavigate = ""
+      if (pathName === "login") routeToNavigate = 1
+      else if (user.role === "Team Member") routeToNavigate = 2
+      else routeToNavigate = 3
+
+      alertMsg({ title: "ÉXITO", text: `${status_code}`, icon: "success" })
+
+      navigate(possibleRoutes[routeToNavigate])
+
+      // console.log("user final -->", user)
+      // console.log("routeToNavigate -->", routeToNavigate)
+      // console.log("possibleRoutes[routeToNavigate] -->", possibleRoutes[routeToNavigate])
+
+      // return () => console.log("desmontando useAuthUSerForm -> efecto")
+    }
+  }, [user, error, status_code, success_request, dispatch, pathName, navigate])
+
+  return { formik, loadingUser, pathName, handleChangeSwitch, handleChangeContinent }
 }
