+++ src/redux/store/reducers/userReducer.js
@@ -16,7 +16,7 @@ export const userReducer = (state = initialUserState, action) => {
       return { ...state, loadingUser: true }
 
     case SUCCESS:
-      console.log("---->>>>> case SUCCESS")
+      // console.log("---->>>>> case SUCCESS")
       const possibleCases = {
         1: { userName: action.payload.userName, role: action.payload.role },
         2: { userName: action.payload.userName, role: action.payload.role, teamID: action.payload.teamID },
@@ -25,16 +25,16 @@ export const userReducer = (state = initialUserState, action) => {
       let dataUser = {}
       if (action.payload.userName || action.payload.role === "Team Member") dataUser = possibleCases[1]
       else if (action.payload.role === "Team Leader") dataUser = possibleCases[2]
-      console.log("state en SUCCESSSSSSSSS", state)
+      // console.log("state en SUCCESSSSSSSSS", state)
       return { loadingUser: false, user: dataUser, error: "", status_code: action.payload.status_code, success_request: true }
 
     case FAILURE:
-      console.log("case FAILURE")
+      // console.log("case FAILURE")
       return { loadingUser: false, user: [], error: action.payload, status_code: "", success_request: false }
 
     case RESET_USER_NOTIFICATION:
-      console.log("---->>>>> case SUCCESS - state", state)
-      console.log("RESET_USER_NOTIFICATION --> RETURNNNN --> ", { ...initialUserState, user: state.user })
+      // console.log("---->>>>> case SUCCESS - state", state)
+      // console.log("RESET_USER_NOTIFICATION --> RETURNNNN --> ", { ...initialUserState, user: state.user })
       return { ...initialUserState, user: state.user }
 
     default:
