+++ src/redux/store/types/types.js
@@ -2,5 +2,5 @@ export const TYPES = {
   REQUEST: "USER_REQUEST",
   SUCCESS: "USER_SUCCESS",
   FAILURE: "USER_FAILURE",
-  RESET_USER_NOTIFICATION: "RESET_USER_NOTIFICATION"
+  RESET_USER_NOTIFICATION: "RESET_USER_NOTIFICATION",
 }
