+++ src/utilities/utilAlert/utilAlertMsg.js
@@ -1,13 +1,6 @@
-import Swal from "sweetalert2";
+import Swal from "sweetalert2"
 
-export const alertMsg = async ({
-  title,
-  text,
-  icon,
-  confirmButtonText = "Aceptar",
-}) => {
-  // console.log("tbn");
-  // console.log("recibo", title, text, confirmButtonText);
+export const alertMsg = async ({ title, text, icon, confirmButtonText = "Aceptar" }) => {
   let takeResponse = await Swal.fire({
     title: `${title}`,
     text: `${text}`,
@@ -16,7 +9,7 @@ export const alertMsg = async ({
     width: "400px",
     timer: 3000,
     timerProgressBar: true,
-  });
+  })
 
-  return takeResponse;
-};
+  return takeResponse
+}
