+++ src/views/Error404/Error404.jsx
@@ -1,3 +1,3 @@
-export default function Error404() {
+export default function Error404({ pageTransition }) {
   return <div className="container">Página no existente</div>
 }
